Fix the installation popup and announce the user about the data collection
